requirements.txt
keras==2.10.0
matplotlib==3.7.1
numpy==1.21.6
pandas==1.5.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
tensorflow==2.10.0